Creating CoreApplication class and moving main logic to it for while
